voice-updated/backend/requirements.txt
flask
flask-cors
python-dotenv
requests
langchain-core
langchain-openai
